refactor(pilot): extract handler functions to handlers.go (GH-1212)
Move issue/task result-processing functions and helpers to a dedicated
handlers.go file:

- handleGitHubIssueWithResult
- handleLinearIssueWithResult
- handleJiraIssueWithResult
- handleAsanaTaskWithResult
- Comment builders: buildExecutionComment, buildFailureComment,
  buildJiraExecutionComment, buildJiraFailureComment,
  buildAsanaExecutionComment, buildAsanaFailureComment,
  formatTokenCountComment
- Helper functions: logGitHubAPIError, parseAutopilotBranch,
  resolveGitHubMemberID, extractGitHubLabelNames

Reduces main.go by ~1,200 lines. All functions remain in package main,
no call sites change.
